Synchronize data stream explicitly
For some reason, the pipe from xml-stream to pelias-dbclient does not
automatically pause at any reasonable data buffering limits. xml-parser
eats all cpu time and exhausts the memory, which first leads to dramatic
slowdown and later a crash. Lots of standard nodejs stream elelements are
being used, and it is hard to say which one does not obey pause/resume
control. Here we fix the error by synchronizing ourselves the parsed data
with the data entering the rest of the pipeline.